Configure travis-ci.org tests
Test Django 1.2 and 1.3; Python 2.6 and 2.7

Don't bother testing Python 2.5 -- it requires an older version of
requests (that doesn't depend on json).

Don't bother testing Django 1.2 -- it requires changes to ``assertRaises``
in the test cases (because the context-manager version of assertRaises is
part of unittest2, which appears in Django 1.3+ or Python 2.7+).

Don't bother testing Django 1.5 (yet).